[#10]feat: 기존 엔티티 수정,패키지 이동
