--- src/main/java/streamingsettlement/streaming/repository/StreamingJpaRepository.java
@@ -1,7 +1,0 @@
-package streamingsettlement.streaming.repository;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import streamingsettlement.streaming.domain.entity.Streaming;
-
-public interface StreamingJpaRepository extends JpaRepository<Streaming,Long> {
-}
