+++ src/main/java/streamingsettlement/streaming/repository/jpa/StreamingAdvertisementJpaRepository.java
@@ -1,4 +1,4 @@
-package streamingsettlement.streaming.repository;
+package streamingsettlement.streaming.repository.jpa;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import streamingsettlement.streaming.domain.entity.StreamingAdvertisement;
@@ -8,4 +8,5 @@ import java.util.Optional;
 
 public interface StreamingAdvertisementJpaRepository extends JpaRepository<StreamingAdvertisement,Long> {
     Optional<StreamingAdvertisement> findByStreamingIdAndPosition(Long streamingId, Integer position);
+    List<StreamingAdvertisement> findByStreamingId(Long streamingId);
 }
